moving app to it's own file under lib/dashing
